tail: use LogReplay
The oplog doesn't have any indexes set. This means that if you have a
query on any field, it will scan through the entirety of the
oplog.

To get around this, mongo has a cursor setting you can change [1] so
that it uses binary search to find the point in the oplog where you want
to start tailing [2].

I figured this out by trying to run this against mongodbraf-dev, which
has a giant (8M records) oplog. While this was running, I ran db.currentOp(). This
showed the tail really struggling (the cursor ran for a few minutes). I then
googled for "mongodb oplog tail ts" and finding this [3] thread.

cc @ericavonb @templaedhel

[1] http://docs.mongodb.org/manual/reference/method/cursor.addOption/#DBQuery.Option.oplogReplay
[2] https://jira.mongodb.org/browse/NODE-94?focusedCommentId=461734&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-461734
[3] https://groups.google.com/forum/#!topic/mongodb-user/E7BSv624nBg